fix: add details to erc20
When I added this custom ERC20 to MetaMask, I had to specify how many
decimals to use. For openzeppelin v3, the default is 18. I
*think* that's the default for v2 as well? But it wasn't clear from the
[docs]. It's best to be explicit, so I added some details.

Based on the initial supply that we mint of 1 billion and the
instructions that instruct to send 1000 to NEAR, it seems like we were
treating this as having 0 decimal places, so that's what I've set here.

  [docs]: https://docs.openzeppelin.com/contracts/2.x/erc20